Split up the TypePack header into individual headers.
