+++ inc/osvr/TypePack/Head.h
@@ -1,0 +1,24 @@
+/** @file
+    @brief Header
+
+    @date 2015
+
+    @author
+    Sensics, Inc.
+    <http://sensics.com/osvr>
+*/
+
+// Copyright 2015 Sensics, Inc.
+// TypePack is part of OSVR-Core.
+//
+// Use, modification and distribution is subject to the
+// Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef INCLUDED_Head_h_GUID_F1F875F3_EA89_45BA_75E4_73A35A8DE89E
+#define INCLUDED_Head_h_GUID_F1F875F3_EA89_45BA_75E4_73A35A8DE89E
+
+#include "SplitList.h"
+
+#endif // INCLUDED_Head_h_GUID_F1F875F3_EA89_45BA_75E4_73A35A8DE89E
