+++ inc/osvr/TypePack/Length.h
@@ -1,0 +1,33 @@
+/** @file
+    @brief Header
+
+    @date 2015
+
+    @author
+    Sensics, Inc.
+    <http://sensics.com/osvr>
+
+    @author
+    Eric Niebler
+*/
+
+// Copyright 2015 Sensics, Inc.
+// TypePack is part of OSVR-Core.
+//
+// Incorporates code from "meta":
+// Copyright Eric Niebler 2014-2015
+//
+// Use, modification and distribution is subject to the
+// Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
+//
+// Project home: https://github.com/ericniebler/meta
+//
+
+#ifndef INCLUDED_Length_h_GUID_7C3F1769_4DC9_479B_92E2_041286FBBCFF
+#define INCLUDED_Length_h_GUID_7C3F1769_4DC9_479B_92E2_041286FBBCFF
+
+#include "Size.h"
+
+#endif // INCLUDED_Length_h_GUID_7C3F1769_4DC9_479B_92E2_041286FBBCFF
