+++ inc/osvr/TypePack/Tail.h
@@ -1,0 +1,24 @@
+/** @file
+    @brief Header
+
+    @date 2015
+
+    @author
+    Sensics, Inc.
+    <http://sensics.com/osvr>
+*/
+
+// Copyright 2015 Sensics, Inc.
+// TypePack is part of OSVR-Core.
+//
+// Use, modification and distribution is subject to the
+// Boost Software License, Version 1.0. (See accompanying
+// file LICENSE_1_0.txt or copy at
+// http://www.boost.org/LICENSE_1_0.txt)
+
+#ifndef INCLUDED_Tail_h_GUID_F49F7DE2_ADEF_4E87_E3F5_C2EF0A6C9672
+#define INCLUDED_Tail_h_GUID_F49F7DE2_ADEF_4E87_E3F5_C2EF0A6C9672
+
+#include "SplitList.h"
+
+#endif // INCLUDED_Tail_h_GUID_F49F7DE2_ADEF_4E87_E3F5_C2EF0A6C9672
